Platformer Testing - Complete

+++ source/game.cpp
@@ -2,7 +2,7 @@
 
 int main(int argc, char* argv[]) {
 	bool flag = false;
-	double accumulator = 0.0;
+	float accumulator = 0.0f;
 
 	GameWorld* gameWorld = new GameWorld();
 	gameWorld->EnableDebugDraw();
